requirements.txt
django==5.0
requests==2.31.0
django-environ==0.11.2